LayoutTests/inspector/console/console-screenshot.html

<!DOCTYPE html>
<html>
<head>
<script src="../../http/tests/inspector/resources/inspector-test.js"></script>
<script>

function createDetachedTest()
{
    let div = document.createElement("div");
    div.id = "detached";
    return div;
}

function createCanvas2DTest() {
    let canvas = document.createElement("canvas");
    canvas.width = 2;
    canvas.height = 2;
    return canvas.getContext("2d");
}

function test()
{
    let suite = InspectorTest.createAsyncSuite("console.screenshot");

    function addTest({name, expression, imageMessageAddedCallback, shouldError}) {
        suite.addTestCase({
            name,
            test(resolve, reject) {
                let listener = WI.consoleManager.addEventListener(WI.ConsoleManager.Event.MessageAdded, async (event) => {
                    let {message} = event.data;

                    let isError = message.level === WI.ConsoleMessage.MessageLevel.Error;
                    if (isError || message.type === WI.ConsoleMessage.MessageType.Image) {
                        WI.consoleManager.removeEventListener(WI.ConsoleManager.Event.MessageAdded, listener);

                        if (isError)
                            InspectorTest.expectThat(shouldError, message.messageText);
                        else
                            InspectorTest.expectEqual(message.type, WI.ConsoleMessage.MessageType.Image, "The added message should be an image.");

                        if (imageMessageAddedCallback)
                            await imageMessageAddedCallback(message);

                        resolve();
                        return;
                    }
                });

                InspectorTest.evaluateInPage(expression)
                .catch(reject);
            },
        });
    }

    addTest({
        name: "console.screenshot.Node.SingleArgument",
        expression: `console.screenshot(document.querySelector("#testNode"))`,
        async imageMessageAddedCallback(message) {
            InspectorTest.expectNotEqual(message.messageText, "data:", "The image should not be empty.");

            let img = await WI.ImageUtilities.promisifyLoad(message.messageText);
            InspectorTest.expectEqual(img.width, 2, "The image width should be 2px.");
            InspectorTest.expectEqual(img.height, 2, "The image height should be 2px.");
        },
    });

    addTest({
        name: "console.screenshot.Node.MultipleArguments",
        expression: `console.screenshot(document.querySelector("#testNode"), "test")`,
        async imageMessageAddedCallback(message) {
            InspectorTest.expectNotEqual(message.messageText, "data:", "The image should not be empty.");

            let img = await WI.ImageUtilities.promisifyLoad(message.messageText);
            InspectorTest.expectEqual(img.width, 2, "The image width should be 2px.");
            InspectorTest.expectEqual(img.height, 2, "The image height should be 2px.");
        },
    });

    addTest({
        name: "console.screenshot.Node.DetachedNonScreenshotable",
        expression: `console.screenshot(createDetachedTest())`,
        shouldError: true,
    });

    addTest({
        name: "console.screenshot.CanvasRenderingContext2D",
        expression: `console.screenshot(createCanvas2DTest())`,
        async imageMessageAddedCallback(message) {
            InspectorTest.expectNotEqual(message.messageText, "data:", "The image should not be empty.");

            let img = await WI.ImageUtilities.promisifyLoad(message.messageText);
            InspectorTest.expectEqual(img.width, 2, "The image width should be 2px.");
            InspectorTest.expectEqual(img.height, 2, "The image height should be 2px.");
        },
    });

    addTest({
        name: "console.screenshot.NonScreenshotableTarget",
        expression: `console.screenshot(42)`,
        async imageMessageAddedCallback(message) {
            InspectorTest.expectNotEqual(message.messageText, "data:", "The image should not be empty.");

            let img = await WI.ImageUtilities.promisifyLoad(message.messageText);
            InspectorTest.expectGreaterThan(img.width, 2, "The image width should be greater than 2px.");
            InspectorTest.expectGreaterThan(img.height, 2, "The image height should be greater than 2px.");
        },
    });

    addTest({
        name: "console.screenshot.NoArguments",
        expression: `console.screenshot()`,
        async imageMessageAddedCallback(message) {
            InspectorTest.expectNotEqual(message.messageText, "data:", "The image should not be empty.");

            let img = await WI.ImageUtilities.promisifyLoad(message.messageText);
            InspectorTest.expectGreaterThan(img.width, 2, "The image width should be greater than 2px.");
            InspectorTest.expectGreaterThan(img.height, 2, "The image height should be greater than 2px.");
        },
    });

    suite.runTestCasesAndFinish();
}
</script>
</head>
<body onload="runTest()">
    <p>Tests for the console.screenshot API.</p>
    <div id="testNode"></div>
    <style>
    #testNode {
        width: 2px;
        height: 2px;
        background-color: red;
    }
    #detached {
        width: 2px;
        height: 2px;
        background-color: blue;
    }
    </style>
</body>
</html>
